Moved Api Client.ts into the collaborative ApiClient file
