+++ ui/src/pages/Home/HomePage.tsx
@@ -4,7 +4,7 @@ const HomePage = () => {
     
     return (
         <div>
-            <h1 className="title">Home Page</h1>
+            <h1 className="title">Whale Watching</h1>
               <Statistics />
         </div>
         );
